Update Covid Exploration Project Queries.sql
In this project, I examine data relating to Covid from the start up to a little over a year into the pandemic in the context of Finland and globally.
I start by exploring Total Cases vs Total deaths to find out the likelihood of dying if a person in Finland contracted Covid in Finland which revealed a low likelihood of dying (about 1%) towards the middle of the pandemic as compared to the beginning. This could have been due to the introduction of vaccines. 
I went on to examining Total Cases vs Total Population to know the percentage of population that contracted covid in Finland. The data revealed that the percentage of population in Finland that contracted covid kept rising steadily and by the April 2021 almost 2% of the population in Finland had contracted covid. 
In analyzing countries with Highest Infection Rate compared to Population, Andorra was among the top 10 as the country with the highest infection rate. Among them were Israel, United States, Montenegro and Slovenia.
On a continent-level comparison, North America and South America had the highest death count with Africa and Oceania having the least. Further analysis revealed that United States, Brazil, Mexico, India and United Kingdom were the countries with the highest death count up to the mid point of the pandemic. 

Looking at the global picture, there was a relatively low daily percentage of death during the pandemic period in question although the data revealed that it was high in the beginning at about 22% but fortunately came down very fast to less than 2% by the April 2021 which would probably be due to the highly effective worldwide vaccination campaign that was launched by the World Health Organization. 
Lastly, turning the focus back to Finland, by the end of April 2021, Finland had vaccinated 1,417,386 people which is roughly 26% of it’s population. An impressive feat by Finland.